generic test generator for pp tests. which uses same test cases as snatch test

+++ tests/test_lib.py
@@ -71,6 +71,7 @@ def setSickbeardGlobals():
     sickbeard.QUALITY_DEFAULT = 4
     sickbeard.SEASON_FOLDERS_DEFAULT = 1
     sickbeard.SEASON_FOLDERS_FORMAT = 'Season %02d'
+    sickbeard.RENAME_EPISODES = True
     
     sickbeard.NAMING_SHOW_NAME = 1
     sickbeard.NAMING_EP_NAME = 1
@@ -208,7 +209,11 @@ def tearDown_test_db():
     if os.path.exists(os.path.join(TESTDIR, TESTCACHEDBNAME)):
         os.remove(os.path.join(TESTDIR, TESTCACHEDBNAME))
 
+
 def setUp_test_episode_file(fileDir=FILEDIR, fileName=FILENAME):
+    if not fileDir is FILEDIR:
+        fileDir = os.path.join(TESTDIR, "complete", fileName)
+
     if not os.path.exists(fileDir):
         os.makedirs(fileDir)
 
@@ -217,15 +222,22 @@ def setUp_test_episode_file(fileDir=FILEDIR, fileName=FILENAME):
     f = open(filePath, "w")
     f.write("foo bar. some fake content")
     f.close()
+    return filePath
 
 
 def tearDown_test_episode_file(fileDir=FILEDIR):
-    shutil.rmtree(FILEDIR)
+    if os.path.isdir(fileDir):
+        shutil.rmtree(FILEDIR)
 
 
 def setUp_test_show_dir(showDir=SHOWDIR):
+    if not showDir is SHOWDIR:
+        showDir = os.path.join(TESTDIR, "shows", showDir)
+
     if not os.path.exists(showDir):
         os.makedirs(showDir)
+
+    return showDir
 
 
 def tearDown_test_show_dir(showDir=SHOWDIR):
